requirements.txt
pydantic==2.10.2
python-dotenv==1.0.1
python-telegram-bot==21.7
requests==2.32.3
stripe==11.3.0